requirements.txt
# AI_pospro — MVP «ИИ для магазина»
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
numpy>=1.24.0
# faiss-cpu>=1.7.4  # опционально; на Windows часто нет wheel — используется numpy fallback
sentence-transformers>=2.2.2
loguru>=0.7.2
pytest>=7.4.0
requests>=2.31.0
